requirements.txt
# === CORE FRAMEWORKS ===
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-core==2.14.0

# === TENSORFLOW ECOSYSTEM ===
tensorflow==2.13.0
keras==2.13.1

# === IMAGE PROCESSING ===
pillow==10.0.1

# === SCIENTIFIC COMPUTING ===
numpy==1.24.3
scipy==1.11.3
pandas==2.1.3

# === SERIALIZATION & MODEL FORMATS ===
h5py==3.9.0
protobuf==4.25.1
joblib==1.3.2

# === HTTP & NETWORKING ===
requests==2.31.0


# === SECURITY & ENCODING ===
cryptography==41.0.7
pyjwt==2.8.0


# === UTILITIES & SYSTEM ===
pyyaml==6.0.1
click==8.1.7
colorama==0.4.6
tqdm==4.66.1

# === RUNTIME & COMPATIBILITY ===
typing-extensions==4.8.0
asgiref==3.7.2
starlette==0.27.0

